Refactor Rocket and RocketEngine
